% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\docType{data}
\name{ReceiverCounter}
\alias{ReceiverCounter}
\alias{$.ReceiverCounter}
\alias{[[.ReceiverCounter}
\title{ReceiverCounter Class}
\format{
An object of class \code{environment} of length 5.
}
\source{
Generated by miniextendr from Rust type \code{ReceiverCounter}
}
\usage{
ReceiverCounter

\method{$}{ReceiverCounter}(self, name)

\method{[[}{ReceiverCounter}(self, name)
}
\description{
Receiver-style counter with \verb{$new()}, \verb{$value()}, \verb{$inc()}, and \verb{$add()} methods.
}
\examples{
rc <- ReceiverCounter$new(1L)
rc$value()
rc$inc()
rc$add(5L)
ReceiverCounter$default_counter()$value()
}
\keyword{datasets}
