% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_altrep_unsafe}
\alias{rpkg_altrep_unsafe}
\alias{unsafe_rpkg_altrep_compact_int}
\alias{unsafe_rpkg_altrep_from_doubles}
\alias{unsafe_rpkg_altrep_from_strings}
\alias{unsafe_rpkg_altrep_from_logicals}
\alias{unsafe_rpkg_altrep_from_raw}
\alias{unsafe_rpkg_altrep_from_list}
\alias{unsafe_rpkg_constant_int}
\alias{unsafe_rpkg_constant_real}
\alias{unsafe_rpkg_simple_vec_int}
\alias{unsafe_rpkg_inferred_vec_real}
\alias{unsafe_rpkg_lazy_int_seq_is_materialized}
\title{ALTREP Unsafe Entry Points}
\source{
Generated by miniextendr from Rust fn \code{rpkg_altrep_compact_int}
}
\usage{
unsafe_rpkg_altrep_compact_int(n, start, step)
}
\description{
ALTREP low-level entry points (unsafe)
}
\examples{
\dontrun{
x <- unsafe_rpkg_altrep_from_doubles(c(1, 2, 3))
unsafe_rpkg_lazy_int_seq_is_materialized(x)
}
}
\keyword{internal}
