% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_panic_tests}
\alias{rpkg_panic_tests}
\alias{add_panic}
\alias{nested_panic}
\alias{add_panic_heap}
\alias{add_r_error}
\alias{add_r_error_heap}
\alias{drop_message_on_success}
\alias{drop_on_panic}
\alias{drop_on_panic_with_move}
\alias{unsafe_C_just_panic}
\alias{unsafe_C_panic_and_catch}
\alias{unsafe_C_r_error}
\alias{unsafe_C_r_error_in_catch}
\alias{unsafe_C_r_error_in_thread}
\alias{unsafe_C_r_print_in_thread}
\title{Panic and Error Handling Tests}
\source{
Generated by miniextendr from Rust fn \code{add_panic}
}
\usage{
add_panic(unused_left, unused_right)
}
\description{
Panic and error handling tests (unsafe)
}
\examples{
try(add_panic(1L, 2L))
try(add_r_error(1L, 2L))
\dontrun{
unsafe_C_r_error_in_thread()
}
}
\keyword{internal}
