% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniextendr_wrappers.R
\name{rpkg_worker_tests}
\alias{rpkg_worker_tests}
\alias{test_worker_return_i32}
\alias{test_worker_return_string}
\alias{test_worker_return_f64}
\alias{test_main_thread_r_api}
\alias{test_main_thread_r_error}
\alias{test_main_thread_r_error_with_drops}
\alias{unsafe_C_worker_drop_on_success}
\alias{unsafe_C_worker_drop_on_panic}
\alias{unsafe_C_test_worker_simple}
\alias{unsafe_C_test_worker_with_r_thread}
\alias{unsafe_C_test_worker_multiple_r_calls}
\alias{unsafe_C_test_worker_panic_simple}
\alias{unsafe_C_test_worker_panic_with_drops}
\alias{unsafe_C_test_worker_panic_in_r_thread}
\alias{unsafe_C_test_worker_panic_in_r_thread_with_drops}
\alias{unsafe_C_test_worker_r_error_in_r_thread}
\alias{unsafe_C_test_worker_r_error_with_drops}
\alias{unsafe_C_test_worker_r_calls_then_error}
\alias{unsafe_C_test_worker_r_calls_then_panic}
\alias{unsafe_C_test_extptr_from_worker}
\alias{unsafe_C_test_multiple_extptrs_from_worker}
\alias{unsafe_C_test_wrong_thread_r_api}
\alias{unsafe_C_test_call_worker_fn_from_main}
\alias{unsafe_C_test_nested_helper_from_worker}
\alias{unsafe_C_test_nested_multiple_helpers}
\alias{unsafe_C_test_nested_with_r_thread}
\alias{unsafe_C_test_nested_worker_calls}
\alias{unsafe_C_test_nested_with_error}
\alias{unsafe_C_test_nested_with_panic}
\alias{unsafe_C_test_deep_with_r_thread_sequence}
\title{Worker Thread Tests}
\source{
Generated by miniextendr from Rust fn \code{test_worker_return_i32}
}
\usage{
test_worker_return_i32()
}
\description{
Worker-thread and main-thread helpers
}
\examples{
test_worker_return_i32()
test_worker_return_string()
test_worker_return_f64()
try(test_main_thread_r_error())
\dontrun{
unsafe_C_test_worker_simple()
}
}
\keyword{internal}
